refactor(vex): introduce new executor and registry for quote provider management
- Deleted `request-key.ts` as it was no longer needed.
- Removed `upstream-routing.ts` and its associated logic.
- Added `executor.ts` to handle quote requests with concurrency limits.
- Created `registry.ts` to manage quote provider instances and metadata.
- Introduced `router.ts` for planning upstream requests based on missed quotes.
- Updated types in `types.ts` to reflect changes in quote handling.
- Enhanced logging and error handling in the new modules.
