+++ .legion/tasks/task-vex-quote-upstream-refactor/context.md
@@ -1,0 +1,62 @@
+# task-vex-quote-upstream-refactor - 上下文
+
+## 会话进展 (2025-12-16)
+
+### ✅ 已完成
+
+- 按你的确认：`terminal_id` / `service_id` 都是全局唯一 UUID，因此 `listWatch((v) => v.serviceInfo.service_id, ...)` 的 watch key 不存在冲突风险。
+- 按你的要求先做小清理：删除临时调试输出（`console.info('11111111', ...)`），并统一 provider discovery 日志 tag 为 `[VEX][Quote]...`；invalid schema 的 ignore 日志也统一为 `[VEX][Quote]...`。
+- 完成 Discovery 梳理要点（供 Design 使用）：
+- 领域能力：L1 路由（prefix 匹配 + field 倒排 + 交集）、按 provider group + max_products_per_request 分批、字段不可用写 `""` 并用 `updated_at` 满足鲜度。
+- 切面能力：服务发现（terminalInfos$ + schema 解析 + group 聚合）、LB（round-robin）、并发治理（同 group 串行 / 全局并发上限）、in-flight 去重（key -> Promise）、日志。
+- 当前不优雅点：`upstream-routing.ts` 顶层创建 `Terminal.fromNodeEnv()` 并 `.subscribe()`（import 即产生副作用，且与 `fillQuoteStateFromUpstream` 传入的 terminal 概念重复）；领域/切面代码交织导致可读性与可测试性差；返回值需要多段解包（route -> map -> plan）。
+- 已处理并闭环你在 `plan.md` 里的 3 条 review：
+- `IQuoteProviderRegistry` 升级为对外 Facade，提供 `planOrThrow/execute/fillQuoteStateFromUpstream`；Router/Executor 改为内部组件；prefix matcher 作为 Router 可替换组件。
+- `unavailableAction` -> `defaultAction`，并移除 `unroutableProducts` 返回值，改为规划阶段直接 `newError` 抛。
+- 目录组织采纳把 `upstream-routing.ts` 迁移为 `apps/virtual-exchange/src/quote/upstream/index.ts`。
+- 完成 upstream-routing 重构：新增 `apps/virtual-exchange/src/quote/upstream/` 目录，将 provider registry / router / executor 分层拆分，并由 Facade `createQuoteProviderRegistry(terminal)` 对外提供 `fillQuoteStateFromUpstream`。
+- 删除 `apps/virtual-exchange/src/quote/upstream-routing.ts`；`apps/virtual-exchange/src/quote/service.ts` 改为创建 registry 并调用 `registry.fillQuoteStateFromUpstream`。
+- 日志 tag 统一为 `[VEX][Quote]...`，并保留关键日志：provider discovery summary / route dispatched / request planned / request received（含 actions JSON）。
+- 运行 `npx prettier -w ...` 完成格式化。
+- 最小编译检查：运行 `./apps/virtual-exchange/node_modules/.bin/tsc -p apps/virtual-exchange/tsconfig.json --noEmit`，目前仓库内存在既有错误（`apps/virtual-exchange/src/position.ts` 缺少字段 `current_price/notional`），与本次改动无关。
+
+### 🟡 进行中
+
+(暂无)
+
+### ⚠️ 阻塞/待定
+
+(暂无)
+
+---
+
+## 关键文件
+
+(暂无)
+
+---
+
+## 关键决策
+
+| 决策                                                                                                                                                                             | 原因                                                                                                                    | 替代方案                                                                                                                                                    | 日期       |
+| -------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- | ----------------------------------------------------------------------------------------------------------------------- | ----------------------------------------------------------------------------------------------------------------------------------------------------------- | ---------- |
+| 沿用 service discovery 的 watch key = `service_id`（因 `service_id` 全局唯一），不把“可能碰撞”作为重构前提；但仍会在重构中把 registry 抽象成接口以隔离 RxJS/TerminalInfos 细节。 | 你确认 `terminal_id`/`service_id` 全局唯一 UUID，不会撞；当前需要解决的是职责分层与可读性/可测试性问题，而非 key 冲突。 | watch key 改为 `encodePath(terminal_id, service_id)` 以防未来约束变化,不用 listWatch，直接每次 terminalInfos$ 变更全量重建 registry（更简单但可能多些计算） | 2025-12-16 |
+
+---
+
+## 快速交接
+
+**下次继续从这里开始：**
+
+1. 联调验证：启动上游 GetQuotes providers + virtual-exchange，调用 `VEX/QueryQuotes` 观察 `[VEX][Quote]...` 日志（UpstreamProviderDiscovery/RouteDispatched/RequestPlanned/RequestReceived），确认路由/分批/LB/in-flight/并发策略符合预期。
+2. 如果需要降低日志噪音：把 `RequestReceived` 的 `JSON.stringify(actions)` 改为仅输出 count 或采样输出（但保留 tag 便于 grep）。
+3. 如果要进一步“更优雅”：考虑把 `listWatch` 订阅的 `.subscribe()` 生命周期集中到 `service.ts`（例如提供 `dispose()`），避免 registry 在单测场景中常驻。
+
+**注意事项：**
+
+- 本次重构完成后：上游路由逻辑入口是 `apps/virtual-exchange/src/quote/upstream/registry.ts#createQuoteProviderRegistry`；`service.ts` 通过 registry 调用 `fillQuoteStateFromUpstream`。
+- 最小 tsc 校验当前被仓库既有错误阻塞：`apps/virtual-exchange/src/position.ts` 缺字段 `current_price/notional`。
+
+---
+
+_最后更新: 2025-12-17 15:07 by Claude_
