+++ .legion/tasks/task-vex-quote-upstream-refactor/tasks.md
@@ -1,0 +1,42 @@
+# task-vex-quote-upstream-refactor - 任务清单
+
+## 快速恢复
+
+**当前阶段**: 阶段 4 - Validation ✅ COMPLETE
+**当前任务**: (none)
+**进度**: 5/5 任务完成
+
+---
+
+## 阶段 1: Discovery ✅ COMPLETE
+
+- [x] 阅读并梳理 `apps/virtual-exchange/src/quote/upstream-routing.ts` 当前的领域逻辑与切面逻辑耦合点，列出需要保留的不变量（L1 路由算法、freshness、in-flight、LB、并发限制、日志）。 | 验收: context.md 记录现状问题清单 + 不变量清单。
+
+---
+
+## 阶段 2: Design ✅ COMPLETE
+
+- [x] 定义分层接口（`IQuoteProviderRegistry` / `IQuoteRouter` / `IGetQuotesExecutor`）以及更易懂的返回值对象（例如 `QuoteUpstreamPlan`），并确定目录拆分方案。 | 验收: plan.md 给出接口草图、职责边界、文件列表与迁移步骤。
+
+---
+
+## 阶段 3: Refactor ✅ COMPLETE
+
+- [x] 移除临时调试残留（例如 `11111111`）并统一日志 tag 为 `[VEX][Quote]`。 | 验收: 不再出现临时 debug 输出；日志 tag 统一。
+- [x] 将 provider discovery 逻辑抽到 registry 模块，实现 `snapshot()`；把路由/分批抽到 router；把 LB/并发/in-flight 抽到 executor；`fillQuoteStateFromUpstream` 只做编排。 | 验收: `upstream-routing.ts` 体积显著下降，职责清晰，调用点无语义变化。
+
+---
+
+## 阶段 4: Validation ✅ COMPLETE
+
+- [x] 运行 prettier + 最小编译检查（尽量只覆盖 apps/virtual-exchange 相关）。 | 验收: 格式无噪音，TypeScript 编译通过（或给出可复现的失败原因）。
+
+---
+
+## 发现的新任务
+
+(暂无)
+
+---
+
+_最后更新: 2025-12-17 15:04_
