+++ apps/virtual-exchange/src/quote/upstream/index.ts
@@ -1,0 +1,2 @@
+export { createQuoteProviderRegistry } from './registry';
+export type { IQuoteProviderRegistry } from './registry';
